adicionando cadastramento e listagem de locações

+++ src/main/java/br/ufscar/dc/dsw/service/impl/LocacaoService.java
@@ -23,7 +23,7 @@ public class LocacaoService implements ILocacaoService {
 	}
 
 	public void excluir(Long id) {
-		dao.deleteById(id);
+		dao.deleteById(id.longValue());
 	}
 
 	@Transactional(readOnly = true)
